# [Enhancement] Use nibble mask in filter_range for ARM