--- a/be/src/column/column_helper.h
+++ b/be/src/column/column_helper.h
@@ -358,6 +358,7 @@ public:
     using ColumnsConstIterator = Columns::const_iterator;
     static bool is_all_const(ColumnsConstIterator const& begin, ColumnsConstIterator const& end);
     static size_t compute_bytes_size(ColumnsConstIterator const& begin, ColumnsConstIterator const& end);
+
     template <typename T, bool avx512f>
     static size_t t_filter_range(const Filter& filter, T* data, size_t from, size_t to) {
         auto start_offset = from;
@@ -432,34 +433,31 @@ public:
             start_offset += kBatchNums;
         }
 #elif defined(__ARM_NEON__) || defined(__aarch64__)
-        const uint8_t* f_data = filter.data() + from;
+        const uint8_t* filter_data = filter.data() + from;
         constexpr size_t data_type_size = sizeof(T);
 
         constexpr size_t kBatchNums = 128 / (8 * sizeof(uint8_t));
         while (start_offset + kBatchNums < to) {
-            uint8x16_t filter = vld1q_u8(f_data);
-            if (vmaxvq_u8(filter) == 0) {
+            const uint8x16_t vfilter = vld1q_u8(filter_data);
+            const uint16x8_t non_zero_mask = vreinterpretq_u16_u8(vmvnq_u8(vceqzq_u8(vfilter)));
+            const uint8x8_t res = vshrn_n_u16(non_zero_mask, 4);
+            uint64_t matches = vget_lane_u64(vreinterpret_u64_u8(res), 0);
+
+            if (matches == 0) {
                 // skip
-            } else if (vminvq_u8(filter)) {
+            } else if (matches == 0xffff'ffff'ffff'ffffull) {
                 memmove(data + result_offset, data + start_offset, kBatchNums * data_type_size);
                 result_offset += kBatchNums;
             } else {
-                for (int i = 0; i < kBatchNums; ++i) {
-                    // the index for vgetq_lane_u8 should be a literal integer
-                    // but in ASAN/DEBUG the loop is unrolled. so we won't call vgetq_lane_u8
-                    // in ASAN/DEBUG
-#if defined(NDEBUG) && !defined(ADDRESS_SANITIZER)
-                    if (vgetq_lane_u8(filter, i)) {
-#else
-                    if (f_data[i]) {
-#endif
-                        *(data + result_offset++) = *(data + start_offset + i);
-                    }
+                matches &= 0x8888'8888'8888'8888ull;
+                for (; matches > 0; matches &= matches - 1) {
+                    uint32_t index = __builtin_ctzll(matches) >> 2;
+                    *(data + result_offset++) = *(data + index);
                 }
             }
 
             start_offset += kBatchNums;
-            f_data += kBatchNums;
+            filter_data += kBatchNums;
         }
 #endif
         // clang-format on