--- a/be/src/formats/orc/apache-orc/c++/include/orc/Common.hh
+++ b/be/src/formats/orc/apache-orc/c++/include/orc/Common.hh
@@ -29,6 +29,8 @@
 
 namespace orc {
 
+using int128_t = __int128;
+
 class FileVersion {
 private:
     uint32_t majorVersion;